chore: separate calibration into runnable (/scripts) and functions (/src)
