+++ src/artefactual/calibration/__init__.py
@@ -3,11 +3,12 @@
 This module exports configuration classes and training functions for calibration.
 """
 
-from artefactual.calibration.helpers.memory import clear_gpu_memory
-from artefactual.calibration.helpers.models import get_model_name, init_llm
 from artefactual.calibration.outputs_entropy import GenerationConfig
 from artefactual.calibration.rates_answers import RatingConfig
 from artefactual.calibration.train_calibration import train_calibration
+from artefactual.calibration.utils.io import load_tqa_from_json, save_to_json
+from artefactual.calibration.utils.memory import clear_gpu_memory
+from artefactual.calibration.utils.models import get_model_name, init_llm
 
 __all__ = [
     "GenerationConfig",
@@ -15,5 +16,7 @@ __all__ = [
     "clear_gpu_memory",
     "get_model_name",
     "init_llm",
+    "load_tqa_from_json",
+    "save_to_json",
     "train_calibration",
 ]
